+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- Disable email verification [#280](https://github.com/rokwire/core-building-block/issues/280)
+
 ### Changed
 - Dissociate permissions from applications [#207](https://github.com/rokwire/core-building-block/issues/207)