+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.0] - 2021-11-01
 ### Added
 - Disable email verification [#280](https://github.com/rokwire/core-building-block/issues/280)
 
@@ -125,7 +126,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)
 - Optimise the Mongo DB collections indexes usage [#146](https://github.com/rokwire/core-building-block/issues/146)
 
-[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/rokwire/core-building-block/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/rokwire/core-building-block/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/rokwire/core-building-block/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/rokwire/core-building-block/compare/v1.5.0...v1.6.0